Refactor LegendBar & Update DeployStateBar
