+++ src/UI/Components/LegendBar/index.ts
@@ -1,1 +1,4 @@
-export * from "./LegendBar";
+export { LegendBar } from "./LegendBar";
+import { Props } from "./Item";
+
+export type LegendItemDetails = Props;
